Add schema validator
